SonOfMan/video/webcam/requirements.txt
flask
flask_cors
flask-socketio
pystun3
aiortc
aiohttp
websockets
python-socketio[asyncio_client]
